requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
httpx==0.28.1
sse-starlette==2.2.1

# Production Server
gunicorn==23.0.0

# Rate Limiting
slowapi==0.1.9

# Redis for caching, rate limiting, request deduplication
redis==5.2.1
aioredis==2.0.1

# LlamaIndex Core
llama-index-core==0.12.10
llama-index-llms-anthropic==0.6.6
llama-index-vector-stores-chroma==0.4.1
llama-index-readers-file==0.4.3

# Embeddings - FastEmbed (lightweight ONNX-based, no PyTorch needed)
# Much smaller than sentence-transformers, uses ~200MB RAM vs 2GB+
llama-index-embeddings-fastembed==0.3.0
fastembed==0.4.2

# Vector Store
chromadb==0.5.23

# Eureka Client
py-eureka-client==0.11.10

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
